+++ providers/Providers.tsx
@@ -6,17 +6,20 @@ import { WagmiProvider } from "./WagmiProvider";
 import { ZoraCreateProvider } from "./ZoraCreateProvider";
 import UserProvider from "./UserProvider";
 import { CrossmintProvider } from "./CrossmintProvider";
+import JamProvider from "./JamProvider";
 
 export function Providers({ children }: { children: ReactNode }) {
   return (
-    <WagmiProvider>
-      <CrossmintProvider>
-        <PrivyProvider>
-          <UserProvider>
-            <ZoraCreateProvider>{children}</ZoraCreateProvider>
-          </UserProvider>
-        </PrivyProvider>
-      </CrossmintProvider>
-    </WagmiProvider>
+    <JamProvider>
+      <WagmiProvider>
+        <CrossmintProvider>
+          <PrivyProvider>
+            <UserProvider>
+              <ZoraCreateProvider>{children}</ZoraCreateProvider>
+            </UserProvider>
+          </PrivyProvider>
+        </CrossmintProvider>
+      </WagmiProvider>
+    </JamProvider>
   );
 }
